+++ src/main/java/com/okbasalman/product_service/adapter/output/repository/ProductRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.okbasalman.product_service.adapter.output.entity.ProductEntity;
 
-public interface ProductRepository extends JpaRepository<ProductEntity, Integer>{
+public interface ProductRepository extends JpaRepository<ProductEntity, Long>{
 
 }
