+++ src/main/java/com/okbasalman/product_service/domain/port/input/ProductUseCase.java
@@ -9,7 +9,7 @@ import com.okbasalman.product_service.domain.model.Product;
 public interface ProductUseCase {
     Product createProduct(ProductCreateDto dto);
     List<Product> getAllProducts();
-    Product getProductById(Integer id);
+    Product getProductById(Long id);
     Product updateProduct(Product product);
-    DeleteProductResultDto deleteProduct(Integer id);
+    DeleteProductResultDto deleteProduct(Long id);
 }
